Polished(?) by ChatGPT
The prompt was:
```
Hi! can you help me to polish the following markdown document which is an installation instruction of my software library "sshcomm" for tclsh?
```

```
Great! Can I have the unrendered markdown of the entire document so I can copy it?
```

Unfortunately, later prompt failed to quote the entire document.